# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Staff Application API
 * OpenAPI spec version: 0.1.0
 */
import type { Application } from './application';

export interface AdminStats {
  totalApplications: number;
  pending: number;
  underReview: number;
  accepted: number;
  denied: number;
  openPositions: number;
  totalPositions: number;
  totalApplicants: number;
  recentApplications: Application[];
}