# lib/api-zod/src/generated/types/position.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Staff Application API
 * OpenAPI spec version: 0.1.0
 */

export interface Position {
  id: number;
  name: string;
  description: string;
  isOpen: boolean;
  /** @nullable */
  discordRoleId: string | null;
  questions: string[];
  applicationCount: number;
  createdAt: string;
}