# lib/api-zod/src/generated/types/positionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Discord Staff Application API
 * OpenAPI spec version: 0.1.0
 */

export interface PositionInput {
  /** @minLength 1 */
  name: string;
  description: string;
  discordRoleId?: string;
  questions?: string[];
  isOpen?: boolean;
}